Added ActionBar, added bminfo, there was something wrong I think

+++ StaffMode Project/src/me/NinetyNine/staff/bminfo/StaffBMInfoInterface.java
@@ -7,11 +7,11 @@ public interface StaffBMInfoInterface {
 
 	void getHistory(Player checker, OfflinePlayer target);
 
-	long getBans(Player checker, OfflinePlayer target);
+	int getBans(Player checker, OfflinePlayer target);
 
-	long getMutes(Player checker, OfflinePlayer target);
+	int getMutes(Player checker, OfflinePlayer target);
 
-	long getKicks(Player checker, OfflinePlayer target);
+	int getKicks(Player checker, OfflinePlayer target);
 
-	long getWarns(Player checker, OfflinePlayer target);
+	int getWarns(Player checker, OfflinePlayer target);
 }
